requirements.txt
huggingface_hub>=0.30.0
pandas>=2.2.0
pyarrow>=15.0.0
PyYAML>=6.0.0
requests>=2.31.0
